+++ patches/src/main/kotlin/dev/jkcarino/revanced/patches/reddit/ad/RemoveAdsAndTelemetryPatch.kt
@@ -7,16 +7,17 @@ import app.revanced.patcher.patch.bytecodePatch
 import com.android.tools.smali.dexlib2.builder.instruction.BuilderInstruction11x
 import dev.jkcarino.revanced.util.proxy
 
-private const val EXTENSION_CLASS_DESCRIPTOR = "Ldev/jkcarino/extension/reddit/AdBlockInterceptor;"
+private const val EXTENSION_CLASS_DESCRIPTOR =
+    "Ldev/jkcarino/extension/reddit/frontpage/AdBlockInterceptor;"
 
 @Suppress("unused")
 val removeAdsAndTelemetryPatch = bytecodePatch(
-    name = "Remove ads, annoyances, and telemetry",
-    description = "Removes ads, annoyances, and telemetry from Home, Popular, Watch, Latest, All, " +
+    name = "Remove ads and telemetry",
+    description = "Removes ads and telemetry from Home, Popular, Watch, Latest, All, " +
         "Custom feeds, Search, and Subreddits, including comments.",
     use = true
 ) {
-    extendWith("extensions/reddit.rve")
+    extendWith("extensions/reddit/frontpage.rve")
 
     compatibleWith("com.reddit.frontpage")